# protocol/bc: use precomputed hash in WitnessHash